refactor(ios): address sixth Copilot review round
- save() now writes the manifest via do/catch and bails out (without
  updating the in-memory list) if encoding/writing fails, so a puzzle
  that won't survive relaunch is never advertised on the home screen;
  the failure trips an assertionFailure in Debug.
- ImageUtilities.thumbnailJPEG takes maxPixel as Int, matching ImageIO's
  kCGImageSourceThumbnailMaxPixelSize semantics.

+++ ios/Pussel/Support/ImageUtilities.swift
@@ -9,7 +9,7 @@ enum ImageUtilities {
     /// ImageIO, which decodes only what the thumbnail needs. Used for the
     /// home-screen puzzle list so a full-size image isn't held in memory or
     /// decoded just to draw a small card.
-    static func thumbnailJPEG(from data: Data, maxPixel: CGFloat = 240) -> Data? {
+    static func thumbnailJPEG(from data: Data, maxPixel: Int = 240) -> Data? {
         guard let source = CGImageSourceCreateWithData(data as CFData, nil) else { return nil }
         let options: [CFString: Any] = [
             kCGImageSourceCreateThumbnailFromImageAlways: true,
